\name{.JavaSigs}
\alias{.JavaSigs}
\non_function{}
\title{A vector of Java type specifiers}
\description{
 This is a named vector containing the pairs
of Java primitive type names (e.g. double, boolean, etc.)
and the corresponding low-level JNI type identifier.
These are things such as
The values ca be used in
in the \code{.sigs} arguments of
\code{\link{.Java}} and 
\code{\link{.JavaConstructor}}.
}
\keyword{Java}