\name{.javaMatchFunctions}
\alias{.javaMatchFunctions}
\non_function{}
\title{Symbolic constants for how classes are matched in conversion}
\description{
The R-Omegahat interface provides 3 built-in routines for
determining whether a Java object matches a particular
class or not. These routines are used in determining 
whether a converter applies to a particular Java object
and is capable of converting it to an R object.
This vector is a named integer vector where the names
are symbolic identifiers for the integers that allow
the R and C code to identify which of these 3 built-in
routines is meant.
}
\keyword{Java}

